requirements.txt
autopep8
certifi
charset-normalizer
click
et-xmlfile
Flask
idna
itsdangerous
Jinja2
MarkupSafe
numpy
openai
openpyxl
pandas
pandas-stubs
pycodestyle
python-dateutil
python-dotenv
pytz
pytest
pytest-cov
vcrpy
testcontainers
respx
requests
six
toml
tqdm
urllib3
Werkzeug
gunicorn
